requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
pydantic
httpx
python-dotenv